add LinearSolvers module
