+++ test/polar_form.jl
@@ -12,7 +12,7 @@ using SparseArrays
 using Test
 using TimerOutputs
 using ExaPF
-import ExaPF: PowerSystem, AD, Precondition, Iterative
+import ExaPF: PowerSystem, AD
 
 const PS = PowerSystem
